test(cli): consolidate build404Hint tests into fetch-diff-gh.test.ts
Team-lead review nit: fewer test files is better. The `build404Hint`
helper lives alongside the gh/HTTP diff-fetch path, so its tests belong
next to the existing `fetchDiffViaGh` tests rather than in a standalone
file. No assertion changes.

Part of #767
