+++ packages/cli/src/__tests__/fetch-diff-gh.test.ts
@@ -1,10 +1,12 @@
 /**
- * Tests for fetchDiffViaGh — the gh CLI diff fetch path.
+ * Tests for the gh/HTTP diff-fetch path in commands/agent.ts.
  *
  * Covers:
- * - Successful gh API call returns diff content
- * - gh CLI failure (not installed, not authenticated) returns null
- * - gh CLI error (non-zero exit) returns null
+ * - `fetchDiffViaGh` — the gh CLI tier (success, ENOENT, non-zero exit,
+ *   HTTP error wrapping, and abort-signal handling).
+ * - `build404Hint` — the 404 hint text used by the HTTP fallback, which
+ *   branches on whether gh is authenticated so we stop blaming auth when
+ *   auth is fine.
  */
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { execFile } from 'node:child_process';
@@ -28,7 +30,7 @@ vi.mock('../tool-executor.js', () => ({
   testCommand: vi.fn(async () => ({ ok: true, elapsedMs: 100 })),
 }));
 
-import { fetchDiffViaGh } from '../commands/agent.js';
+import { fetchDiffViaGh, build404Hint } from '../commands/agent.js';
 
 const mockedExecFile = vi.mocked(execFile);
 
@@ -147,5 +149,36 @@ describe('fetchDiffViaGh', () => {
 
     expect(result).toBeNull();
     expect(killFn).toHaveBeenCalled();
+  });
+});
+
+describe('build404Hint', () => {
+  it('returns the authenticated-branch hint when gh auth passes', () => {
+    const hint = build404Hint(true);
+
+    // Leads with a period+space so it appends cleanly to the base error message.
+    expect(hint.startsWith('. ')).toBe(true);
+    // Authenticated branch must NOT suggest installing or logging into gh —
+    // the whole point of this issue is to stop blaming auth when auth is fine.
+    expect(hint).not.toMatch(/gh auth login/);
+    expect(hint).not.toMatch(/ensure gh CLI is installed/);
+    // It should enumerate the real plausible causes.
+    expect(hint).toMatch(/PR not found/);
+    expect(hint).toMatch(/transient GitHub outage/i);
+    // And point operators at the underlying gh-api warning if there was one.
+    expect(hint).toMatch(/\[fetchDiffViaGh\]/);
+  });
+
+  it('returns the install/login hint when gh auth fails or gh is missing', () => {
+    const hint = build404Hint(false);
+
+    expect(hint.startsWith('. ')).toBe(true);
+    // Keeps the original actionable guidance for users without gh.
+    expect(hint).toMatch(/gh CLI is installed and authenticated/);
+    expect(hint).toMatch(/gh auth login/);
+  });
+
+  it('produces distinct text for the two branches', () => {
+    expect(build404Hint(true)).not.toEqual(build404Hint(false));
   });
 });
